requirements.txt
streamlit
tensorflow
pillow
h5py